refactor: VisitFragment 내 visit을 moment로 변경
